requirements.txt
anthropic>=0.30.0
opencv-python>=4.8.0
Pillow>=10.0.0
streamlit>=1.35.0
jinja2>=3.1.0
